chore: Moved RQ logic only used in `useHasNewUpdate` into that file
