--- mobile/src/features/setting/api/release.ts
@@ -1,41 +1,0 @@
-import { useQuery } from "@tanstack/react-query";
-
-import { settingKeys } from "./_queryKeys";
-
-const RELEASE_NOTES_LINK =
-  "https://api.github.com/repos/MissingCore/Music/releases";
-
-type ReleaseNotes =
-  | { releaseNotes: undefined; version: undefined }
-  | { releaseNotes: string; version: string };
-
-function formatGitHubRelease(data: any): ReleaseNotes {
-  return {
-    version: data.tag_name,
-    // Remove markdown comments w/ regex.
-    releaseNotes: data.body
-      ? data.body.replace(/<!--[\s\S]*?(?:-->)/g, "")
-      : undefined,
-  } satisfies ReleaseNotes;
-}
-
-export async function getLatestRelease() {
-  return {
-    latestStable: await fetch(`${RELEASE_NOTES_LINK}/latest`)
-      .then((res) => res.json())
-      .then((data) => formatGitHubRelease(data)),
-    latestRelease: await fetch(`${RELEASE_NOTES_LINK}?per_page=1`)
-      .then((res) => res.json())
-      .then(([data]) => formatGitHubRelease(data)),
-  };
-}
-
-/** Returns the latest releases. */
-export const useLatestRelease = () =>
-  useQuery({
-    queryKey: settingKeys.release(),
-    queryFn: getLatestRelease,
-    staleTime: Infinity,
-    gcTime: Infinity,
-    retry: false,
-  });
