+++ mobile/src/hooks/useHasNewUpdate.ts
@@ -1,4 +1,6 @@
-import { useLatestRelease } from "@/features/setting/api/release";
+import { useQuery } from "@tanstack/react-query";
+
+import { settingKeys } from "@/constants/QueryKeys";
 
 import { APP_VERSION } from "@/constants/Config";
 
@@ -26,3 +28,43 @@ export function useHasNewUpdate(): UpdateResult {
     return { hasNewUpdate: true, release: usedRelease };
   }
 }
+
+//#region Data
+const RELEASE_NOTES_LINK =
+  "https://api.github.com/repos/MissingCore/Music/releases";
+
+type ReleaseNotes =
+  | { releaseNotes: undefined; version: undefined }
+  | { releaseNotes: string; version: string };
+
+function formatGitHubRelease(data: any): ReleaseNotes {
+  return {
+    version: data.tag_name,
+    // Remove markdown comments w/ regex.
+    releaseNotes: data.body
+      ? data.body.replace(/<!--[\s\S]*?(?:-->)/g, "")
+      : undefined,
+  } satisfies ReleaseNotes;
+}
+
+async function getLatestRelease() {
+  return {
+    latestStable: await fetch(`${RELEASE_NOTES_LINK}/latest`)
+      .then((res) => res.json())
+      .then((data) => formatGitHubRelease(data)),
+    latestRelease: await fetch(`${RELEASE_NOTES_LINK}?per_page=1`)
+      .then((res) => res.json())
+      .then(([data]) => formatGitHubRelease(data)),
+  };
+}
+
+/** Returns the latest releases. */
+const useLatestRelease = () =>
+  useQuery({
+    queryKey: settingKeys.release(),
+    queryFn: getLatestRelease,
+    staleTime: Infinity,
+    gcTime: Infinity,
+    retry: false,
+  });
+//#endregion
